Updated README with note on unpacked extensions
Updated the README after someone kindly pointed me to the fact that .crx files can be unpacked by WinRAR/7-Zip etc, which is probably an easier of recovering the original files to drag into the browser than saving each one from the source folder. Also fixes a JSON problem that seems to occur if you download the manifest from the source folder.